thick MerkleTree trait
